feat(map): indicators computation per year
